fix(types): :bug: don't assume `SinglePartBook` won't have a series; combine redundant ApiBook types
After some testing, it seems Audible is inconsistent with how books in a series are labeled. To combat this, we will now allow single and multi part books to contain a series/publication. We can also remove BookSchema, because chapterInfo doesn't seem to ever be attached to a book, so not sure why this was created. Also, since series are optional, we can put those into the main book type

fixes #589